+++ packages/snap-account-service/src/index.ts
@@ -12,8 +12,12 @@ export type {
 export type {
   SnapAccountServiceEnsureMigratedAction,
   SnapAccountServiceEnsureReadyAction,
+  SnapAccountServiceGetAccountAssetsAction,
+  SnapAccountServiceGetAccountBalancesAction,
+  SnapAccountServiceGetAccountTransactionsAction,
   SnapAccountServiceGetSnapsAction,
   SnapAccountServiceHandleKeyringSnapMessageAction,
+  SnapAccountServiceResolveAccountAddressAction,
 } from './SnapAccountService-method-action-types';
 export { SnapPlatformWatcher } from './SnapPlatformWatcher';
 export type { SnapPlatformWatcherConfig } from './SnapPlatformWatcher';